move compose file latest hash logic to quickstart vesioning
