+++ metadata-ingestion/src/datahub/cli/docker_cli.py
@@ -797,59 +797,9 @@ def quickstart(
     )
 
 
-def is_it_a_version(version: str) -> bool:
-    """
-    Checks if a string is a valid version.
-    :param version: The string to check.
-    :return: True if the string is a valid version, False otherwise.
-    """
-    return re.match(r"v?\d+\.\d+(\.\d+)?", version) is not None
-
-
-def parse_version(version: str) -> Tuple[int, int, int, int]:
-    """
-    Parses a version string into a tuple of integers.
-    :param version: The version string to parse.
-    :return: A tuple of integers representing the version.
-    """
-    version = re.sub(r"v", "", version)
-    parsed_version = tuple(map(int, version.split(".")))
-    # pad with zeros if necessary
-    if len(parsed_version) == 2:
-        parsed_version = parsed_version + (0, 0)
-    elif len(parsed_version) == 3:
-        parsed_version = parsed_version + (0,)
-    return parsed_version
-
-
-def compare_versions(
-    version1: Tuple[int, int, int, int], version2: Tuple[int, int, int, int]
-) -> bool:
-    """
-    Compares two versions.
-    :return: True if version1 is greater than version2, False otherwise.
-    """
-    for i in range(4):
-        if version1[i] > version2[i]:
-            return True
-        elif version1[i] < version2[i]:
-            return False
-    return False
-
-
 def get_docker_compose_base_url(version_tag: str) -> str:
     if os.environ.get("DOCKER_COMPOSE_BASE"):
         return os.environ["DOCKER_COMPOSE_BASE"]
-
-    # new CLI version is downloading the composefile corresponding to the requested version
-    # if the version is older than v0.10.1, it doesn't contain the setup job labels and the
-    # the checks will fail, so in those cases we pick the composefile from v0.10.1 which contains
-    # the setup job labels
-    if is_it_a_version(version_tag):
-        if compare_versions(parse_version("v0.10.0"), parse_version(version_tag)):
-            # The merge commit where the labels were added
-            # https://github.com/datahub-project/datahub/pull/7473
-            version_tag = "e9f060dae399ffc98dfbb121d77b3cbefef88417"
 
     return f"https://raw.githubusercontent.com/datahub-project/datahub/{version_tag}"
